add --dry-run option

+++ delete.go
@@ -11,8 +11,12 @@ type deleteOperation struct {
 	path string
 }
 
-func (o deleteOperation) Perform(ctx context.Context) error {
+func (o deleteOperation) Perform(ctx context.Context, dryRun bool) error {
 	log.Printf("NUKE %s", fp.Base(o.path))
+	if dryRun {
+		return nil
+	}
+
 	return os.RemoveAll(o.path)
 }
